+++ app/Domains/Auth/Services/UserService.php
@@ -109,7 +109,10 @@ class UserService extends BaseService
             ]);
 
             $user->syncRoles($data['roles'] ?? []);
-            $user->syncPermissions($data['permissions'] ?? []);
+
+            if (!config('boilerplate.access.user.only_roles')) {
+                $user->syncPermissions($data['permissions'] ?? []);
+            }
         } catch (Exception $e) {
             DB::rollBack();
 
@@ -149,7 +152,10 @@ class UserService extends BaseService
             if (! $user->isMasterAdmin()) {
                 // Replace selected roles/permissions
                 $user->syncRoles($data['roles'] ?? []);
-                $user->syncPermissions($data['permissions'] ?? []);
+
+                if (!config('boilerplate.access.user.only_roles')) {
+                    $user->syncPermissions($data['permissions'] ?? []);
+                }
             }
         } catch (Exception $e) {
             DB::rollBack();