+++ resources/views/backend/auth/user/create.blade.php
@@ -108,7 +108,10 @@
                     </div>
 
                     @include('backend.auth.includes.roles')
-                    @include('backend.auth.includes.permissions')
+
+                    @if (!config('boilerplate.access.user.only_roles'))
+                        @include('backend.auth.includes.permissions')
+                    @endif
                 </div>
             </x-slot>